+++ lib/spacejam.dart
@@ -1,10 +1,14 @@
 export "src/theme/text_styles.dart";
+export "src/theme/theme.dart";
 
 // Widgets
+export "src/page/spacejam_page.dart";
 export "src/collection/collection_item.dart";
-export "src/contentbox/content_box.dart";
+export "src/container/container_child.dart";
+export "src/container/container.dart";
 export "src/collection/collection.dart";
-export "src/imagepage/image_page.dart";
+export "src/image_page/image_page.dart";
 export "src/appbar/appbar_action.dart";
 export "src/appbar/appbar.dart";
 export "src/button/button.dart";
+export "src/image_box/image_box.dart";