+++ lib/src/collection/collection.dart
@@ -9,20 +9,25 @@ import "../private/min.dart";
 import "../theme/text_styles.dart";
 
 /// Display a bunch of clickable options with or without a section title.
-class Collection extends StatelessWidget {
+class SpaceJamCollection extends StatelessWidget {
   /// Constructor
-  const Collection({required this.items, Key? key, this.title})
+  const SpaceJamCollection({required this.items, Key? key, this.title})
       : super(key: key);
 
-  /// List of [CollectionItem]s to display.
-  final List<CollectionItem> items;
+  /// List of [SpaceJamCollectionItem]s to display.
+  final List<SpaceJamCollectionItem> items;
 
   /// Title of the collection
   final String? title;
 
   @override
   Widget build(BuildContext context) => items.isNotEmpty
-      ? Column(
+      ? Padding(
+    padding: EdgeInsets.only(bottom: (MediaQuery.of(context).size.width +
+        MediaQuery.of(context).size.height) /
+        2 *
+        .04,),
+    child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: <Widget>[
             title != null
@@ -46,6 +51,6 @@ class Collection extends StatelessWidget {
               children: items,
             ),
           ],
-        )
+        ),)
       : const Min();
 }